Introduce EditionBase parent contract
The primary goal is to maximize shared code between Edition and SingleBatchEdition.

To do that we clarify interfaces:
- IEditionBase includes all the behavior that is expected of an edition (initializer, metadata, size limits, time limits, operator filters...)
- IRealTimeMintable includes the traditional minting and sale methods
- IBatchMintable includes the batch-specific minters and getters

We can then compose them as follows:
- an IEdition is an IEditionBase + IRealTimeMintable
- an IBatchEdition is an IEditionBase + IBatchMintable

We also introduce an abstract base class EditionBase that handles most of the edition behavior, except for the minting parts.

Benefits, both regular Editions, SingleBatchEditions and the upcoming MultiBatchEditions now have the same:
- initializer
- private compact state
- getters/setters for the time limit, edition size, approved minters, ...
- metadata functions
- operator filter
- ERC2981 support